Split into Modules
Move the contents of the main file into a set of submodules.
