+++ test/test_data.py
@@ -24,149 +24,7 @@
 from nose.tools import *
 
 # Module to test
-from parsegen import *
-
-class TestParsegen(object):
-	"""Test Parsegen
-	
-	Tests the `parsegen` module.
-	"""
-	
-	def test_errors(self):
-		e_str = "dummy string"
-		
-		e = ParseError(e_str)
-		assert str(e).startswith('parsegen:')
-		assert 'parse error' in str(e)
-		assert e_str in str(e)
-		
-		e = GrammarError(e_str)
-		assert str(e).startswith('parsegen:')
-		assert 'grammar error' in str(e)
-		assert e_str in str(e)
-
-	
-	def test_parse_buffer(self):
-		
-		res = parse_buffer(" %% %% ")
-		assert res != None
-		assert res[2] == " "
-		
-		h, e, c = parse_buffer(
-		"""
-		TOKEN = Tok_TOKEN
-		FUZZBAZ
-		%language = c
-		%lexer_entry = Lex_getNextToken();
-		%%
-		# empty grammar section
-		%%
-		user_code and stuff
-		"""
-		)
-		
-		assert type(h) == Header
-		assert type(c) == str
-		assert type(e) == dict
-		
-		assert h.options["language"] == "c"
-		assert h.options["lexer_entry"] == "Lex_getNextToken();"
-		
-		assert h.terminals['TOKEN'] == 'Tok_TOKEN'
-		assert 'TOKEN' in h.terminals
-		assert 'FUZZBAZ' in h.terminals
-		
-		assert len(e) == 0
-		
-		assert c == "\n\t\tuser_code and stuff\n\t\t"
-		
-	def test_write_grammar(self):
-		header, expansions, user_code = parse_buffer("""
-		
-		WORLD
-		
-		%language = C
-		%prefix = yy
-		%lexer_function = Lex_getNextToken()
-		%token_type = Lex_Token
-
-		%%
-		
-		main := hello
-		
-		hello := WORLD
-		
-		%%
-		
-		hello world
-		
-		""")
-		
-		write_grammar(header, expansions, user_code, sys.stdout)
-		
-		assert False
-
-	def test_parse_buffer_errors(self):
-		
-		assert_raises(ParseError, lambda: parse_buffer(" %% %% %% "))
-		assert_raises(ParseError, lambda: parse_buffer(" %% "))
-	
-	def test_expansion_creation(self):
-		_, exps, _ = parse_buffer(
-		"""
-		%language = c
-		FOO = Tok_FOO
-		BAZ = Tok_BAZ
-		BAR = Tok_BAR
-		%%
-		main := baz bar_prime BAZ
-		bar  := FOO BAR
-		bar_prime := bar bar_prime
-		bar_prime := 
-		baz := bar_prime
-		baz := BAZ
-		%%
-		// this is a comment
-		"""
-		)
-		
-		assert len(exps) == 4
-		assert len(exps['main']) == 1
-		assert len(exps['bar']) == 1
-		assert len(exps['bar_prime']) == 2
-		assert len(exps['baz']) == 2
-		
-		main = exps['main']
-		bar = exps['bar']
-		bar_prime = exps['bar_prime']
-		baz = exps['baz']
-		
-		# Check the nullability
-		assert not main.is_nullable()
-		assert not bar.is_nullable()
-		assert bar_prime.is_nullable()
-		assert baz.is_nullable()
-		
-		# Check the first sets
-		print(main.first, bar.first, bar_prime.first, baz.first)
-		assert main.first == {'FOO', 'BAZ'}
-		assert bar.first == {'FOO'}
-		assert bar_prime.first == {'FOO'}
-		assert baz.first == {'BAZ', 'FOO'}
-		
-		# Check the follow sets
-		print(main.follow, bar.follow, bar_prime.follow, baz.follow)
-		assert main.follow == set()
-	
-	def test_invalid_expansions(self):
-		
-		assert_raises(
-			GrammarError, lambda: parse_buffer(" %% main := TOKEN %% "))
-		assert_raises(
-			GrammarError, lambda: parse_buffer(" %% main := fdass %% "))
-		assert_raises(
-			GrammarError, lambda: parse_buffer(
-				"TOKEN %% TOKEN := invalid \n invalid := %% "))
+from parsegen.data import *
 
 class TestHeader(object):
 	"""Test Header
@@ -282,5 +140,4 @@ class TestSymbol(object):
 		s.add_follow(['bar', 'baz'])
 		
 		assert s.follow == {'foo', 'bar', 'baz'}
-
