+++ test/test_errors.py
@@ -1,0 +1,46 @@
+# This file is part of Parsegen and is licensed as follows:
+#
+# Copyright (c) 2012 Will Speak
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+# Test helpers
+from nose.tools import *
+
+# Package to test
+from parsegen.errors import *
+
+class TestErrors(object):
+	"""Test Errors
+	
+	Tests the exceptions that can be thrown by the module.
+	"""
+	
+	def test_errors(self):
+		e_str = "dummy string"
+		
+		e = ParseError(e_str)
+		assert str(e).startswith('parsegen:')
+		assert 'parse error' in str(e)
+		assert e_str in str(e)
+		
+		e = GrammarError(e_str)
+		assert str(e).startswith('parsegen:')
+		assert 'grammar error' in str(e)
+		assert e_str in str(e)
